Creates tableView and populates data into tableView; MVP;

+++ obj-c-lab-1/obj-c-lab-1/Queue.m
@@ -29,7 +29,7 @@ NSMutableArray *queue;
 }
 
 -(int) count {
-    return queue.count;
+    return (int)queue.count;
 }
 
 -(void) printQueue {
